# frontend/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CampusEats v1",
    "description": "单校区点餐演示；所有支付退款均为模拟",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/admin/products/{type}/{id}": {
      "put": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/employees/{id}": {
      "put": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "updateEmployee",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Employee"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/categories/{id}": {
      "put": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "updateCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "deleteCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/products/{type}": {
      "get": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "products",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProductView"
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "product",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/employees": {
      "get": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "employees",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EmployeeView"
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "employee",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Employee"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/categories": {
      "get": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "adminCategories",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CategoryView"
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "category",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/user/menu/items": {
      "get": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "menu",
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProductView"
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/reports": {
      "get": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "reports",
        "parameters": [
          {
            "name": "begin",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/reports/export": {
      "get": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "export",
        "parameters": [
          {
            "name": "begin",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "string",
                      "format": "byte"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/dashboard": {
      "get": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "dashboard",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/couriers": {
      "get": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "couriers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": {}
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Component": {
        "type": "object",
        "properties": {
          "dishId": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "copies": {
            "type": "integer",
            "format": "int32",
            "maximum": 50,
            "minimum": 1
          }
        },
        "required": [
          "dishId"
        ]
      },
      "Flavor": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 32,
            "minLength": 0
          },
          "values": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 32,
              "minLength": 0
            },
            "minItems": 1
          }
        },
        "required": [
          "name",
          "values"
        ]
      },
      "Product": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 32,
            "minLength": 0
          },
          "categoryId": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "price": {
            "type": "string",
            "example": "12.00"
          },
          "enabled": {
            "type": "boolean"
          },
          "description": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "image": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "flavors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Flavor"
            }
          },
          "components": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Component"
            }
          }
        },
        "required": [
          "categoryId",
          "name",
          "price"
        ]
      },
      "Employee": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 32,
            "minLength": 0
          },
          "name": {
            "type": "string",
            "maxLength": 32,
            "minLength": 0
          },
          "role": {
            "type": "string",
            "minLength": 1,
            "pattern": "ADMIN|OPERATOR|DELIVERER"
          },
          "enabled": {
            "type": "boolean"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12
          }
        },
        "required": [
          "name",
          "role",
          "username"
        ]
      },
      "Category": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 32,
            "minLength": 0
          },
          "type": {
            "type": "integer",
            "format": "int32",
            "maximum": 2,
            "minimum": 1
          },
          "sort": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "enabled": {
            "type": "boolean"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "ComponentView": {
        "type": "object",
        "properties": {
          "dishId": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "copies": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "currentName": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "FlavorView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "name": {
            "type": "string"
          },
          "values": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ProductView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "name": {
            "type": "string"
          },
          "categoryId": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "price": {
            "type": "string",
            "example": "12.00"
          },
          "image": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "remaining": {
            "type": "integer",
            "format": "int32"
          },
          "itemType": {
            "type": "string"
          },
          "flavors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FlavorView"
            }
          },
          "components": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComponentView"
            }
          },
          "defaultQuota": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CategoryView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "EmployeeView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "mustChangePassword": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}